# frontend/api/scripts/outputs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CortexDJ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sessions/{session_id}/similar-tracks": {
      "get": {
        "tags": [
          "retrieval"
        ],
        "summary": "Get Similar Tracks",
        "description": "Return the top-k tracks whose CLAP audio embedding is closest to the\nsession's EEG embedding in the joint contrastive space.",
        "operationId": "getSimilarTracks",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarTracksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "List Sessions",
        "description": "List all EEG sessions.",
        "operationId": "listSessions",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionListResponse": {
        "properties": {
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionSchema"
            },
            "type": "array",
            "title": "Sessions"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "sessions",
          "total"
        ],
        "title": "SessionListResponse"
      },
      "SessionSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "participant_id": {
            "type": "string",
            "title": "Participant Id"
          },
          "dataset_source": {
            "type": "string",
            "title": "Dataset Source"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          },
          "metadata_extra": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Extra"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "participant_id",
          "dataset_source",
          "recorded_at",
          "duration_seconds",
          "created_at"
        ],
        "title": "SessionSchema"
      },
      "SimilarTrackSchema": {
        "properties": {
          "spotify_id": {
            "type": "string",
            "title": "Spotify Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "itunes_preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Itunes Preview Url"
          },
          "similarity": {
            "type": "number",
            "title": "Similarity"
          }
        },
        "type": "object",
        "required": [
          "spotify_id",
          "title",
          "artist",
          "similarity"
        ],
        "title": "SimilarTrackSchema"
      },
      "SimilarTracksResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/SimilarTrackSchema"
            },
            "type": "array",
            "title": "Tracks"
          },
          "k": {
            "type": "integer",
            "title": "K"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "tracks",
          "k"
        ],
        "title": "SimilarTracksResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}